Elasticsearch searches default text field and analyzed text field.
Split Postgres and Elasticsearch into separate source files.
